Add VaultOverview endpoint

+++ src/Domain/Entities/Deployment/CatalogDeployment.cs
@@ -3,7 +3,7 @@ public class CatalogDeployment
 {
     public required string Address { get; init; }
 
-    public required ulong ContractChainId { get; init; }
+    public required ushort ContractChainId { get; init; }
 
     public virtual ChainDeployment? Chain { get; init; } //Navigation Property
     public virtual List<FeatureDeployment>? Features { get; set; } //Navigation Property
