+++ src/Domain/Entities/Deployment/FeatureDeployment.cs
@@ -6,7 +6,7 @@ public class FeatureDeployment
     public required string Name { get; init; }
     public required string Description { get; init; }
 
-    public required ulong ContractChainId { get; init; }
+    public required ushort ContractChainId { get; init; }
     public required string CatalogAddress { get; init; }
 
     public required string FeeTokenSymbol { get; init; }
